Adding tests and refactoring

+++ lib/form526_backup_submission/service.rb
@@ -32,13 +32,11 @@ module Form526BackupSubmission
         file.pdf_path
       when CarrierWave::SanitizedFile
         file.file
+      when Hash
+        get_file_path_from_objs(file[:file])
       else
         file
       end
-    end
-
-    def get_file_from_obj(obj)
-      obj.is_a?(Hash) ? obj[:file] : obj.file
     end
 
     def generate_tmp_metadata(metadata)
@@ -48,20 +46,19 @@ module Form526BackupSubmission
       json_tmpfile
     end
 
-    def get_file_name(file)
-      file_with_full_path = get_file_path_from_objs(file)
-      File.basename(file_with_full_path)
-    end
-
-    def upload_logging_logic(file_with_full_path:, attachments:)
-      if file_with_full_path =~ /tmp/
-        if Rails.env.production?
-          File.delete(file_with_full_path)
-          attachments.each(&:delete)
-        else
-          Rails.logger.info("Would have deleted file #{file_with_full_path} if in production env.")
-          attachments.each do |evidence_file|
-            to_del = get_file_from_obj(evidence_file)
+    def upload_deletion_logic(file_with_full_path:, attachments:)
+      if Rails.env.production?
+        File.delete(file_with_full_path) if File.exist?(file_with_full_path)
+        attachments.each do |evidence_file|
+          to_del = get_file_path_from_objs(evidence_file)
+          # dont delete the instructions pdf we keep on the fs and send along for bdd claims
+          File.delete(to_del) if File.exist?(to_del) && (!to_del =~ /bdd.pdf/)
+        end
+      else
+        Rails.logger.info("Would have deleted file #{file_with_full_path} if in production env.")
+        attachments.each do |evidence_file|
+          to_del = get_file_path_from_objs(evidence_file)
+          if File.exist?(to_del) && !to_del.include?('bdd_instructions.pdf')
             Rails.logger.info("Would have deleted file #{to_del} if in production env.")
           end
         end
@@ -70,10 +67,10 @@ module Form526BackupSubmission
 
     def upload_doc(upload_url:, file:, metadata:, attachments: [])
       json_tmpfile = generate_tmp_metadata(metadata)
-      file_with_full_path = get_file_name(file)
-
+      file_with_full_path = get_file_path_from_objs(file)
+      file_name = File.basename(file_with_full_path)
       params = { metadata: Faraday::UploadIO.new(json_tmpfile.path, Mime[:json].to_s, 'metadata.json'),
-                 content: Faraday::UploadIO.new(file_with_full_path, Mime[:pdf].to_s, file_name) }
+                 content: Faraday::UploadIO.new(file, Mime[:pdf].to_s, file_name) }
       attachments.each.with_index do |attachment, i|
         file_path = get_file_path_from_objs(attachment[:file])
         file_name = attachment[:file_name]
@@ -82,7 +79,11 @@ module Form526BackupSubmission
 
       response = perform :put, upload_url, params, { 'Content-Type' => 'multipart/form-data' }
 
-      upload_logging_logic(file_with_full_path: file_with_full_path, evidence: evidence)
+      # dont delete bdd or other static docs
+
+      raise response.body unless response.success?
+
+      upload_deletion_logic(file_with_full_path: file_with_full_path, attachments: attachments)
 
       response
     ensure
